requirements.txt
aiortc>=1.6.0
openai>=1.0.0
pyaudio>=0.2.11
numpy>=1.24.0
websockets>=11.0.0
python-dotenv>=1.0.0
requests>=2.31.0
keyboard>=1.13.0